Added useState hooks regarding sorting the items in asc or desc order, min and max price filtering.
Search -> useEffect: Added a timeout constant, which value depends on if the user tries to sort or is typing in an input field.

Search -> useEffect: Added pull out from the api of 'vehicleType', 'pricePerDay' and 'presentmentCurr'.

Search -> useEffect: Added checks for wheter the user is using the input fields for min and max price for filtering.

Search -> useEffect: Added checks for sorting the 'formattedData' if the user clicked on asc. order or on desc. order filter button.

Added the rendering of a new component 'PriceFilter', which takes care of the min and max price input fields.

Added the rendering of a new component 'ListItemsOrder', which takes care of the items ordering (asc. or desc. order).